+++ apps/landing/index.html
@@ -2,7 +2,11 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/png" href="/src/shared/assets/pabicon.webp" />
+    <link
+      rel="icon"
+      type="image/png"
+      href="/src/shared/assets/pabicon.webp"
+    />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <meta property="og:title" content="kareer" />
     <meta
