+++ apps/landing/src/layout/footer/landing-footer.tsx
