+++ apps/landing/src/pages/landing/landing-page.tsx
@@ -1,15 +1,13 @@
-import {
-  CareerRoadmapSection,
-  EarlyAccessSection,
-  HeroSection,
-  JobRecommendationSection,
-  LandingFooter,
-  LandingHeader,
-  RealVoicesSection,
-  StatsSection,
-  TodoManagementSection,
-  WhyKareerSection,
-} from '@widgets/landing';
+import LandingFooter from '@layout/footer/landing-footer';
+import LandingHeader from '@layout/header/landing-header';
+import CareerRoadmapSection from '@sections/career-roadmap/career-roadmap-section';
+import EarlyAccessSection from '@sections/early-access/early-access-section';
+import HeroSection from '@sections/hero/hero-section';
+import JobRecommendationSection from '@sections/job-recommendation/job-recommendation-section';
+import RealVoicesSection from '@sections/real-voices/real-voices-section';
+import StatsSection from '@sections/stats/stats-section';
+import TodoManagementSection from '@sections/todo-management/todo-management-section';
+import WhyKareerSection from '@sections/why-kareer/why-kareer-section';
 
 import * as styles from './landing-page.css';
