+++ apps/landing/src/sections/hero/pain-point-item/pain-point-item.css.ts
