+++ apps/landing/src/sections/job-recommendation/job-preview-card/job-preview-card.css.ts
