+++ apps/landing/src/sections/job-recommendation/job-recommendation-preview/job-recommendation-preview.css.ts
