+++ apps/landing/src/sections/real-voices/voice-card/voice-card.css.ts
