+++ apps/landing/src/shared/components/feature-section/feature-section.tsx
