+++ apps/landing/src/shared/constants/hero.ts
