+++ apps/landing/src/shared/constants/index.ts
