+++ apps/landing/src/shared/constants/job-preview.ts
