+++ apps/landing/src/shared/constants/navigation.ts
