--- apps/landing/src/shared/ui/index.ts
@@ -1,1 +1,0 @@
-export { default as Badge } from './badge/badge';
