+++ apps/landing/tsconfig.json
@@ -5,8 +5,10 @@
     "types": ["vite/client", "vite-plugin-svgr/client"],
     "paths": {
       "@app/*": ["src/app/*"],
-      "@pages/*": ["src/pages/*"],
-      "@widgets/*": ["src/widgets/*"],
+      "@landing": ["src/pages/landing/index.ts"],
+      "@landing/*": ["src/pages/landing/*"],
+      "@layout/*": ["src/layout/*"],
+      "@sections/*": ["src/sections/*"],
       "@shared/*": ["src/shared/*"]
     }
   },
